replace file when updating

+++ backend/controllers/Photographer.js
@@ -1,4 +1,5 @@
 const Photographer = require('../models/Photographer')
+const fs = require('fs')
 
 exports.createPhotographer = (req, res, next) => {
   const thingObject = JSON.parse(req.body.thing)
@@ -25,17 +26,48 @@ exports.getAllPhotographer = (req, res, next) => {
 }
 
 exports.modifyOnePhotographer = (req, res, next) => {
-  const photographerObject = req.file ? {
-    ...JSON.parse(req.body.thing),
-    portraitUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-  } : { ...req.body }
-  Photographer.updateOne({ _id: req.params.id },{...photographerObject, _id:req.params.id})
-    .then(() => res.status(200).json({ message: 'Photographer modified !' }))
-  .catch(error=> res.status(400).json({error}))
+  if (req.file) {
+    const photographerObject = {
+      ...JSON.parse(req.body.thing),
+      portraitUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    }
+    Photographer.findOne({ _id: req.params.id })
+      .then(photographer => {
+        const filename = photographer.portraitUrl.split('/images/')[ 1 ]
+        fs.unlink(`images/${filename}`, () => {
+          Photographer.updateOne({ _id: req.params.id }, { ...photographerObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Photographer modified !' }))
+            .catch(error => res.status(400).json({ error }))
+        })
+      })
+      .catch (error => res.status(500).json({ error }))
+  } else {
+    const photographerObject = { ...req.body }
+    Photographer.updateOne({ _id: req.params.id },{...photographerObject, _id:req.params.id})
+    .then(() => res.status(200).json({message:'Photographer modified !'}))
+    .catch(error=> res.status(400).json({error}))
+  }
 }
 
 exports.deleteOnePhotographer = (req, res, next) => {
-  Photographer.deleteOne({ _id: req.params.id })
-  .then(() => res.status(200).json({message:'Photographer deleted !'}))
-  .catch(error=> res.status(400).json({error}))
+  Photographer.findOne({ _id: req.params.id })
+    .then(photographer => {
+      if (!photographer) {
+        return res.status(404).json({
+          error: new Error('Photographer not found')
+        })
+      }
+      if (photographer.userId !== req.auth.userId) {
+        return res.status(401).json({
+          error: new Error('request non authorized')
+        })
+      }
+      const filename = photographer.portraitUrl.split('/images/')[1]
+      fs.unlink(`images/${filename}`, () => {
+        Photographer.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({message: 'Photographer deleted !'}))
+          .catch((error) => res.status(400).json({ error }))
+      })
+    })
+    .catch(error => res.status(500).json({ error }))
 }
